requirements.txt
Flask=3.0.3
mySql-connector-python==9.0.0